Explicitely disable ssl in mariadb server config
After mariadb 10.5.18, mariadb fails to start if ssl cert parameters
are configured in the file but the certificates do not exist [1]. This
problem was fixed in puppetlabs-mysql 13.1.0[2] but we use older
versions (13.0.0 in stable/zed for example) in stable branch jobs.

CentOS Stream 9 has just bumped to mariadb-10.5.22 and this is breaking
all non-ssl jobs.

I'm also fixing u-c url to use the proper version for Xena and pinning
Ruby gems versions required for CentOS Stream 8.

[1] https://mariadb.com/kb/en/mariadb-10-5-18-release-notes/#ssl
[2] https://github.com/puppetlabs/puppetlabs-mysql/issues/1509

Change-Id: I317c5b1fa8967dc40437fa6e3258b7d8f2c79e58