Generate Puppetfile
We want to generate our Puppefile automatically
based on the value set in openstack_modules.txt and
external_modules.txt files.
A bot will then propose the correct version based on
puppet-openstack-integration project for the common
puppet modules. For the modules only required by
Packstack, the bot will propose the latest version.
Note we can pin by tag or commit if needed.

Some notes:
- We move camptocamp/puppet-systemd to voxpupuli namespace
to avoid the redirection (i.e camptocamp is voxpupuli now)
- We keep the pinning for nssdb, remote and sysctl as there is
no update since a while now and we don't want to get version
behing the current commit. We might unpin if there is a new
version.
- We remove usage of puppet-ssh requirement. It was used by
puppet-tripleo [1] which is now retired. There is no other
usage of it, so we can remove the requirement.
- We move from arioch/puppet-redis to voxpupuli/puppet-redis
which is used by p-o-i [2]. Note arioch/puppet-redis has been
deprecated since Mar 15th 2019 [3] and author has suggested
voxpupuli one as replacement.

[1] https://review.opendev.org/c/x/packstack/+/912479
[2] https://github.com/openstack/puppet-openstack-integration/blob/b61132232dba85b18b2563480ae21400f4e58774/external_modules.txt#L19
[3] https://forge.puppet.com/modules/arioch/redis/readme

Change-Id: I2e67c1c7e48b92485066a2e9c1ce4bd5fe602b04