Redis: Properly disable unix socket
undef is silently translated to the default value by Puppet, thus
the socket was not disabled as intended.

Recent puppet-redis accpets '' as  the indicator to disable unix domain
socket so we can use it instead. Also drop the _perm option which is
not actually used when unix domain socket is disabled.

Change-Id: I4a8b476f0573f5fb771174ab86fad293462df060